requirements.txt
Flask==2.0.3
flask-redis==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-SSE==1.0.0
PyJWT==2.4.0
PyMySQL==1.0.2
pyTelegramBotAPI==4.7.0
requests==2.26.0
SQLAlchemy==1.4.37
flask_cors